--- packages/modal/src/components/FullWindowOverlay/FullWindowOverlay.ios.ts
@@ -1,1 +1,0 @@
-export { FullWindowOverlay } from "react-native-screens";
