+++ packages/modal/src/constants/defaultConfig.ts
@@ -1,0 +1,15 @@
+import { ANIMATION_DURATION_IN_MS } from "./animations";
+import { ModalProps } from "./types";
+
+export const defaultConfig: ModalProps = {
+  animationInTiming: ANIMATION_DURATION_IN_MS,
+  animationOutTiming: ANIMATION_DURATION_IN_MS,
+  hideBackdrop: false,
+  backdropColor: "rgba(0, 0, 0, 0.5)",
+  dampingFactor: 0.2,
+  direction: "bottom",
+  swipeVelocityThreshold: 500,
+  onBackButtonPress: undefined,
+  onBackdropPress: undefined,
+  style: {},
+} satisfies ModalProps;
